requirements.txt
requests
beautifulsoup4
lxml
striprtf
langchain
langchain-community
scikit-learn
spacy
scipy
nltk
tk
python-dotenv
pymupdf
keybert
textblob
langdetect
readability-lxml
cython
lxml_html_clean
time
openai
tiktoken
chromadb
transformers
sentence-transformers
